requirements.txt
pyrofork
dnspython
TgCrypto-pyrofork
pymongo
motor
python-dotenv
apscheduler
uvloop
loguru
flask
gunicorn
requests
